backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.0.0
aiofiles==23.2.1
gunicorn==21.2.0

